chore(versions.tf): add Terraform configuration file to manage provider versions and remote state storage
The `versions.tf` file is added to the `localstack` directory. This file is used to define the required provider versions for Terraform. In this case, the `aws` provider is set to version 5.17.0, the `github` provider is set to version 5.37.0, and the `random` provider is set to version 3.3.1.

Additionally, the `aws` provider is configured with the region set to "us-east-1", and the `github` provider is configured with the owner set to "liquibase".

No newline at the end of the file.