chore(main.yml): add installation of Liquibase SDK snapshot to the build process
The Liquibase SDK snapshot is now being installed during the build process. This allows for the usage of the Liquibase SDK in subsequent steps. The installation is done using the Liquibase branch specified in the `configure-build` step, along with the repositories `liquibase/liquibase` and `liquibase/liquibase-pro`. Additionally, the workflow IDs `build.yml` and `run-tests.yml` are provided for the Pro and non-Pro workflows respectively. The installed snapshot version is then displayed in the build overview.