chore(pom.xml): add liquibaseHome and allowInstall configuration to the maven-scm-plugin
The liquibaseHome configuration is added to specify the location of the Liquibase binary. This allows the plugin to find and execute Liquibase commands.

The allowInstall configuration is added to enable the installation of Liquibase if it is not found in the specified liquibaseHome location. This ensures that Liquibase is available for the plugin to use.