chore(main.yml): add steps to set liquibase-core version to 0-SNAPSHOT and install snapshot CLI
The changes in this commit add two new steps to the main.yml workflow file.

The first step uses the `mvn versions:set-property` command to set the `liquibase-core.version` property to `0-SNAPSHOT`. This is done to update the version of the liquibase-core library used in the project.

The second step uses the `mvn liquibase-sdk:install-snapshot-cli` command to install the snapshot CLI for the liquibase-sdk. This step also includes additional parameters to specify the branch search order and the repositories to search for the liquibase-sdk.

These changes are made to ensure that the liquibase-core version is set to the desired value and to install the necessary tools for working with the liquibase-sdk.